Changed the following in jimmy.py
PEP 8 style guide checker (now called pycodestyle) is installed.
The settings.json file has been updated with the new settings to
enable control against the PEP 8 standard.

The entire code has been adjusted to comply with the PEP 8 standard.

The activate() method in the TestBoxPrinter class has been modified
to allow printing the text "Activate POWER contact."
instead of the text "Activate contact IN 0 (POWER)."
when the user is prompted to activate the power switch.

In function action_1(), Input.POWER is changed to Input(0) to
conform with the rest of the code.