Modularization of interactive testprogram and adjustment for spaces in the output from printfunction under the class TestBoxPrinter
The entire code has so far been collected in the jimmy.py module

In connection with the modularization, the module jimmy.py has been deleted
and the code has been distributed as follows in the following modules:

Module: __init__.py

	- There is nothing in this module.
	  The module just needs to be there to tell the python
	  that this is not a nomal directory, but actually a package of modules.

Module: __main__.py

    - This module should act as the main program (main file),
	  understood in the sense that it is the one that is run when writing
	  the command (python -m PCD), where -m means run, in the terminal.

Module: actions.py

    - All 113 actions are gathered in this module

Module: domain.py

    - In this module my ENUM classes Input and Output are added.

Module: printer.py

    - In this module, the actions on my test equipment
	  are gathered in the class TestBoxPrinter.

      At the bottom of the module, it is ensured that the module
	  can both be tested separately and used to import from:

      if __name__ == "__main__" :)

      The module is run separately by typing the command (python -m PCD),
	  where -m means run, in the terminal.

In the module printer.py I have added the variable activation,
which I use to assess whether there should be spaces
before I print the next thing to be done in the current action.
It's just to make a presentable formatting when printed on the screen.

If, for example, 2 inputs must be activated one directly after the other,
it must not make spaces between the lines, but if the next action changes
from something on input to something on output and vice versa, then it must make spaces
between the lines.