In __main__.py i'm trying to implement a menu via the curses library, because I don't need a GUI based program at this point.
The curses library supplies a terminal-independent screen-painting and keyboard-handling facility
for text-based terminals.

I've googled for things about curses and tried different things, but eventually I
don't know how to integrate it with my code.

Curses also behaves differently depending on which IDE I'm using.
In pycharm I can go up/down via the up/down arrows, but in VS Code I can't move the up/down at all.

Also in VS Code there seems to be errors when using curses, because of pylint
e.g. pylint reports that Module 'curses' has no 'color_pair' member even though it can see it via . notation.
And that's true for other functions as well.

I tried out another package called 'curses-menu' which is a simple Python menu-based system on the terminal using curses.

With this package I can make a menu without having errors in VS Code and here i'm able to go up/down, but only via choosing the number in the menu.
In my submenu I have more than 10 entries, so I can't choose all of them.

curses-menu seems to be the right solution.

So the conclusion is that i'm know using pycharm and continues with the curses-menu.

In action.py I want to have only one dictionary containing all my actions, grouped under the correct sequence which correponds to a door funtion in the menu.
It should look like this.

sequence_dictionary
{
    sequence_1: {"action start_position" : action start_position, "action 1" : action 1, "action 2" : action 2,
                 "action 3" : action 3, "action 4" : action 4, "action 5" : action 5, "action 6" : action 6,
				 "action 7" : action 7 , "action 8" : action 8 , "action 9" : action 9 , "action 10" : action 10,
				 "action 11" : action 11, "action 12" : action 12}
}

sequence_dictionary is a default dict with sequence_1 as key and ordered dict with all the actions belonging to sequence_1, as value.

Right now nothing works and the files are chaotic, because i can't get all this to work. And therefore the modules __main__.py, actions.py and printer.py look like a mess
with different experiments.