Changed code __main__
Update Organization / structure in README with description og domain.py and picture of statemachine

Corrected code, so that the program stops the test and exits to the command line interpreter if any errors occur, which
means that the repairman answers with a no if one or more outputs is not as expected.
This was only true for all other actions than action 1, so this adjustment also counts for action 1.